# backend/docs/openapi.yml
openapi: 3.1.0
info:
  title: Orchestrate Studio API
  description: |
    Internal Developer Platform (IDP) API for managing applications, workflows, and executions.
    
    ## Authentication
    All API endpoints (except `/health`, `/ready`, `/live`) require Bearer token authentication via Keycloak JWT.
    
    Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Multi-Tenancy
    - All resources are scoped to the user's organization (`org_id`)
    - Requests automatically filter results by the user's org context
    - Cross-org access is prevented at the database level
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@orchestrate.dev
    url: https://orchestrate.dev/support
  license:
    name: Proprietary
    url: https://orchestrate.dev/license
  x-logo:
    url: 'https://orchestrate.dev/logo.png'
    altText: Orchestrate Studio Logo
servers:
  - url: http://localhost:3000
    description: Local Development Server
  - url: https://api.orchestrate.dev
    description: Production Server
  - url: https://staging-api.orchestrate.dev
    description: Staging Server
...
paths:
...
  /api/auth/user:
    get:
      summary: Get current authenticated user
      tags:
        - Authentication
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/verify:
    get:
      summary: Verify JWT token validity
      tags:
        - Authentication
      operationId: verifyToken
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Token invalid or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/profile:
    get:
      summary: Get user profile details
      tags:
        - Authentication
      operationId: getUserProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  profile:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/auth/logout:
    post:
      summary: Logout user session
      tags:
        - Authentication
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /api/auth/signup:
    post:
      summary: Create new user account
      tags:
        - Authentication
      operationId: signupUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  userId:
                    type: string
                  message:
                    type: string
        '400':
          description: Invalid signup data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/applications:
    get:
      summary: List all applications in org
      tags:
        - Applications
      operationId: listApplications
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by application status
          schema:
            type: string
            enum: [active, inactive, archived]
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          description: Result pagination offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Applications retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  total:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Application'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      summary: Create new application
      tags:
        - Applications
      operationId: createApplication
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
      responses:
        '201':
          description: Application created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Application'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak JWT bearer token
  schemas:
...
    User:
      type: object
      description: Authenticated user information from JWT token
      properties:
        id:
          type: string
          description: User ID (sub claim from JWT)
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          description: User email address
          example: user@orchestrate.dev
        name:
          type: string
          description: User full name
          example: John Doe
        preferred_username:
          type: string
          description: Username (from preferred_username claim)
          example: john.doe
        org_id:
          type: string
          description: Organization ID for multi-tenant isolation
          example: org-acme-corp
    UserProfile:
      type: object
      description: Extended user profile information
      properties:
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: user@orchestrate.dev
        name:
          type: string
          example: John Doe
        username:
          type: string
          example: john.doe
    SignupRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - password
        - confirmPassword
      properties:
        firstName:
          type: string
          description: User first name
          minLength: 1
          maxLength: 100
          example: John
        lastName:
          type: string
          description: User last name
          minLength: 1
          maxLength: 100
          example: Doe
        email:
          type: string
          format: email
          description: User email (must be unique)
          example: john.doe@example.com
        password:
          type: string
          format: password
          description: Account password (minimum 8 characters, 1 uppercase, 1 number, 1 special char)
          minLength: 8
          example: SecurePass123!
        confirmPassword:
          type: string
          format: password
          description: Password confirmation (must match password)
          example: SecurePass123!

    # ========================================================================
    # Application Schema
    # ========================================================================
    Application:
      type: object
      description: |
        Represents an application within the platform.
        Applications are containers for workflows and are scoped to an organization.
      required:
        - id
        - org_id
        - name
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique application identifier
          example: app-prod-platform-01
        org_id:
          type: string
          description: Organization ID (multi-tenant isolation)
          example: org-acme-corp
        name:
          type: string
          description: Application name (unique within org)
          minLength: 1
          maxLength: 255
          example: Production Platform
        description:
          type: string
          description: Application description
          maxLength: 1000
          nullable: true
          example: Main application for production deployments
        status:
          type: string
          description: Application lifecycle status
          enum: [active, inactive, archived]
          example: active
        metadata:
          type: object
          description: Custom metadata and tags
          nullable: true
          additionalProperties: true
          example:
            owner: platform-team
            costCenter: ENG-001
            environment: production
        created_at:
          type: string
          format: date-time
          description: RFC 3339 timestamp when application was created
          example: '2024-01-15T10:30:00Z'
        updated_at:
          type: string
          format: date-time
          description: RFC 3339 timestamp of last update
          example: '2024-01-20T14:15:00Z'
    CreateApplicationRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Application name
          minLength: 1
          maxLength: 255
          example: My Production App
        description:
          type: string
          description: Optional application description
          maxLength: 1000
          nullable: true
          example: Handles production workloads
        status:
          type: string
          description: Initial application status (defaults to 'active')
          enum: [active, inactive, archived]
          default: active
...
    ErrorResponse:
      type: object
      description: |
        Standard error response format for all API errors.
        All error responses use this consistent structure.
      required:
        - success
        - error
        - message
      properties:
        success:
          type: boolean
          description: Always false for error responses
          example: false
        error:
          type: object
          required:
            - code
            - type
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - INVALID_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - INTERNAL_ERROR
                - SERVICE_UNAVAILABLE
              example: UNAUTHORIZED
            type:
              type: string
              description: Error category
              enum:
                - authentication
                - authorization
                - validation
                - not_found
                - conflict
                - business_logic
                - system
              example: authentication
            message:
              type: string
              description: Human-readable error message
              example: Invalid or expired JWT token
        message:
          type: string
          description: Detailed error description
          example: The provided JWT token has expired. Please obtain a new token from the authentication endpoint.
        trace_id:
          type: string
          description: Unique request trace ID for support debugging
          example: 1705756200000-7a8b9c
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: When error occurred
          example: '2024-01-20T14:15:00Z'
        details:
          type: object
          description: Additional error context (varies by error type)
          additionalProperties: true
          nullable: true
          example:
            field: email
            expected: valid email format
            received: invalid-email
  responses:
...
    UnauthorizedError:
      description: Unauthorized - missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              type: authentication
              message: Missing or invalid JWT token
            message: Please provide a valid Bearer token in the Authorization header
            trace_id: 1705756200000-abc123
            timestamp: '2024-01-20T14:15:00Z'
...
    BadRequestError:
      description: Bad request - invalid input or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: INVALID_REQUEST
              type: validation
              message: Validation failed
            message: Request validation failed
            details:
              field: name
              error: required field
            trace_id: 1705756200000-abc123
            timestamp: '2024-01-20T14:15:00Z'
...